+++ policy_simulation.h
@@ -14,8 +14,8 @@ struct list_entry {
 
 struct task_stats_entry {
 	struct task_key key;
-	int hits;
-	int misses;
+	unsigned long hits;
+	unsigned long misses;
 	UT_hash_handle hh;
 };
 
@@ -24,14 +24,14 @@ struct policy_simulation {
 	struct task_stats_entry *task_stats;
 	void (*hit_update)(struct policy_simulation *, struct list_entry *);
 	void (*miss_update)(struct policy_simulation *, unsigned long);
-	int hits;
-	int misses;
+	unsigned long hits;
+	unsigned long misses;
 };
 
 
 struct policy_simulation *policy_simulation_init(void (*hit_update)(struct policy_simulation *, struct list_entry *), void (*miss_update)(struct policy_simulation *, unsigned long));
 void policy_simulation_track_access(struct policy_simulation *ps, const struct event *e);
-void policy_simulation_evict(struct policy_simulation *ps, int num_to_evict);
+void policy_simulation_evict(struct policy_simulation *ps, unsigned long num_to_evict);
 int policy_simulation_size(struct policy_simulation *ps);
 void policy_simulation_print(struct policy_simulation *ps);
 void mru_hit_update(struct policy_simulation *ps, struct list_entry *hit_entry);
